+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.15.6
+
+* added equality checking and membership test to Optional